Backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
Pillow==10.0.0
stripe==6.0.0
python-dotenv==1.0.0
